+++ praw/endpoints.py
@@ -184,6 +184,7 @@ API_PATH = {
     "upload_image":            "r/{subreddit}/api/upload_sr_img",
     "user":                    "user/{user}/",
     "user_about":              "user/{user}/about/",
+    "user_by_fullname":        "/api/user_data_by_account_ids",
     "user_flair":              "r/{subreddit}/api/user_flair_v2",
     "users_new":               "users/new",
     "users_popular":           "users/popular",