+++ tests/integration/models/reddit/test_redditor.py
@@ -9,6 +9,7 @@ from ... import IntegrationTest
 
 class TestRedditor(IntegrationTest):
     FRIEND = "PyAPITestUser3"
+    FRIEND_FULLNAME = "t2_6c1xj"
 
     @mock.patch("time.sleep", return_value=None)
     def test_block(self, _):
@@ -40,6 +41,12 @@ class TestRedditor(IntegrationTest):
             assert "date" in redditor.__dict__
             assert "created_utc" not in redditor.__dict__
             assert hasattr(redditor, "created_utc")
+
+    def test_fullname_init(self):
+        self.reddit.read_only = False
+        with self.recorder.use_cassette("TestRedditor.test_fullname_init"):
+            redditor = self.reddit.redditor(fullname=self.FRIEND_FULLNAME)
+            assert redditor.name == self.FRIEND
 
     def test_gild__no_creddits(self):
         self.reddit.read_only = False